Set blank to empty string if calling rule with same param twice
avoid a segfault where we try to assign blanks that haven't been set,
e.g. calling a two-param macro from a one-pattern rule: when we try to
setup for the call, it tries to fill the myblanks array with npar-1=1
elements, but blanks is empty.

  <section-def-macros>
    <def-macro n="foo" npar="2"> <let><var n="meh"/><lit v="aa"/></let> </def-macro>
  </section-def-macros>

  <section-rules>
    <rule>
      <pattern> <pattern-item n="ndef"/> </pattern>
      <action>
        <call-macro n="foo"><with-param pos="1"/><with-param pos="1"/></call-macro>
        <out> <chunk> <clip pos="1" part="whole"/> </chunk> </out>
      </action>
    </rule>
  </section-rules>

similarly with when there are two patterns, but you call the macro
with the last pattern twice (no following blanks)