Pipe-wrapper to insert/remove [<STREAMCMD:SETVAR:${AP_SETVAR}>]
For reading by cg-proc.

This is inserted at start of stream and after every NUL (since tf uses
NUL's in document translation). But these wrappers are not meant to be
kept open when pipelines are kept open in Apy – they should be
restarted so that the env-var can be changed per request :-)