Precompile def-cats
This shifts rule transducer growth from exponential to linear, causing
a very slight increase in overhead for very small files (whose
compilation time was negligible to begin with) and having an
increasing effect as the number of rules and particularly the length
of the patterns increases.

| file         | old      | new     | improvement |
|--------------|----------|---------|-------------|
| eng-spa.t1x  |   147 MB | 16.5 MB |          9x |
|              |   2.87 s |   0.25s |         11x |
| nob-nno.t1x  |   155 MB | 16.7 MB |          9x |
|              |  3.267 s | 0.081 s |         40x |
| oci-fra.t2bx |  2655 MB | 15.7 KB |        168x |
|              | 310.98 s | 0.045 s |       6909x |

The downside of this is that we can no longer report when rules using
different def-cats have overlapping paths, though this can be checked
by apertium-lint, so it's not a huge loss.