Allow inverting preference values with initial minus
If preferences.xml has

  <preference id="ise_ize">
    <description lang="eng">Use -ize instead of -ise</description>
    <options>
      <option id="agonise"/>
      <option id="idolise"/>
      <option id="exorcise"/>
      <option id="lionise"/>
    </options>

and user does

    export AP_SETVAR="ise_ize=-agonise:idolise,color_colour"

then that will now be equivalent to

    export AP_SETVAR="ise_ize=exorcise:lionise,color_colour"

Very handy when list of options is a bit long and user wants to turn
off the preference for a single word.

Uses the preferences xml from "$DATADIR/prefs/$PAIR.xml" (if it can't
find that, tries to fall back by removing everything after last dash
in mode name if it can't find that, so debug modes should be supported
as well).